Implement get exemptions service method
