+++ src/main/java/uk/gov/companieshouse/exemptions/ServiceUnavailableException.java
@@ -1,6 +1,8 @@
 package uk.gov.companieshouse.exemptions;
 
-public class ServiceUnavailableException extends RuntimeException {
+import org.springframework.dao.DataAccessException;
+
+public class ServiceUnavailableException extends DataAccessException {
     public ServiceUnavailableException(String message) {
         super(message);
     }
